Fix nbsp problem and it tests moved to unit tests
Some pages contain the non breaking space char inserted as unicode char.
This is not great as then we have some words separated by something that
is not a regular space. This fix just replaces this char with a regular space.

Also, downloaded all the urls we were using in our IT tests so they can be more
of "unit tests" and be executed in much less time.